This one too.

+++ bin/upgrading/update-db-general.pl
@@ -4328,7 +4328,7 @@ EOF
         do_sql( 'UPDATE renames SET status="U" WHERE renuserid = 0' );
     }
 
-    if ( !check_dbnote( 'remove_countries_from_codes' ) ) {
+    if ( table_relevant("codes") && !check_dbnote( 'remove_countries_from_codes' ) ) {
         do_sql( 'DELETE FROM codes WHERE type = "country"' );
         set_dbnote( 'remove_countries_from_codes', 1 );
     }
